backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
anthropic==0.28.0
sentence-transformers==3.0.0
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.7.0
python-dotenv==1.0.1
groq==0.9.0
